Refactor creation function
